[#29] chore: 클로즈 훅 공통화
