+++ src/shared/ui/Menu/SideMenu/SideMenu.tsx
@@ -8,7 +8,7 @@ import {
 } from "@/shared/ui/Menu/SideMenu/type/categoryType";
 import SideMenuTab from "@/shared/ui/Menu/SideMenu/SideMenuTab";
 import useSideMenuStore from "@/stores/useSideMenuStore";
-import useCloseSideMenu from "@/shared/ui/Menu/SideMenu/hooks/useCloseSideMenu";
+import useClose from "@/shared/@common/hooks/useClose";
 
 /**
  * 메인 페이지의 layout.tsx에서 API 요청해서 카테고리 데이터 받아오고 layout.tsx에서 사용하면됨
@@ -27,10 +27,10 @@ const SideMenu = ({ categories }: SideMenuProps) => {
   // SideMenu 영역 지정
   const sideMenuRef = useRef<HTMLDivElement>(null);
   // SideMenu 닫힘 버튼 구현 대신 외부영역 및 ESC로 닫는 훅 호출
-  useCloseSideMenu(isOpenSideMenu, setIsOpenSideMenu, sideMenuRef);
+  useClose(isOpenSideMenu, setIsOpenSideMenu, sideMenuRef);
 
   const commonClass = "py-[25px] px-[10px] h-full bg-black-1C";
-  const openClass = `${commonClass} mobile:relative mobile:z-10 mobile:w-[180px] md:w-[180px] lg:w-[220px]`;
+  const openClass = `${commonClass} mobile:absolute mobile:z-10 mobile:w-[180px] md:w-[180px] lg:w-[220px]`;
   const closeClass = `${commonClass} hidden md:block lg:block md:w-[180px] lg:w-[220px]`;
 
   return (
